Support overriding example in action definition
The `example` method of an action takes a block. If there is an :ok
response and that has a media type, this block will be called during doc
generation with the rendered example. It must return a valid example
for that media type, but can edit the given example as it sees fit.

Signed-off-by: Sean McGivern <sean.mcgivern@rightscale.com>